+++ cg/apps/tb/api.py
@@ -18,8 +18,8 @@ LOG = logging.getLogger(__name__)
 class TrailblazerAPI:
     """Interface to Trailblazer for `cg`."""
 
-    __STARTED_STATUSES = ["completed", "failed", "pending", "running"]
-    __ONGOING_STATUSES = ["pending", "running"]
+    __STARTED_STATUSES = ["completed", "failed", "pending", "running", "error"]
+    __ONGOING_STATUSES = ["pending", "running", "error"]
 
     def __init__(self, config: dict):
         self.service_account = config["trailblazer"]["service_account"]