+++ cg/meta/workflow/mip.py
@@ -12,7 +12,7 @@ from cg.apps.lims import LimsAPI
 from cg.apps.mip import parse_trending
 from cg.apps.mip.base import MipAPI
 from cg.apps.mip.confighandler import ConfigHandler
-from cg.apps.scoutapi import ScoutAPI
+from cg.apps.scout.scoutapi import ScoutAPI
 from cg.apps.tb import TrailblazerAPI
 from cg.apps.tb.models import TrailblazerAnalysis
 from cg.constants import (
@@ -432,7 +432,7 @@ class MipAnalysisAPI(ConfigHandler, MipAPI):
         return self.db.family(case_id)
 
     def get_analyses_to_clean(self, before: dt.datetime) -> list:
-        analyses_to_clean = self.db.analyses_to_clean(pipeline="mip", before=before)
+        analyses_to_clean = self.db.analyses_to_clean(pipeline=Pipeline.MIP_DNA, before=before)
         return analyses_to_clean.all()
 
     def get_slurm_job_ids_path(self, case_id: str) -> Path: