+++ tests/store/api/test_store_add.py
@@ -119,6 +119,7 @@ def test_add_pool(store: Store):
         category="rml",
         description="Ready-made",
         percent_kth=80,
+        percent_reads_guaranteed=75,
         sequencing_depth=0,
     )
     store.add_commit(application)